tset: split food test
